Modernize Airliner World + USA for Xcode 26 / iOS 26
Revive the 2010 cocos2d 0.99.4 games so they build and run (fully playable)
on Xcode 26 / the iOS 26 simulator on Apple Silicon.

Build/project:
- ARCHS_STANDARD_32_BIT -> ARCHS_STANDARD, SDKROOT iphoneos4.0/3.2 -> iphoneos,
  add IPHONEOS_DEPLOYMENT_TARGET, drop empty GCC_VERSION
- Remove the discontinued iAd framework and all banner code
- Info.plist: lock orientation, hide status bar, UIRequiresFullScreen

cocos2d 0.99.4 fixes for 64-bit / modern iOS:
- intptr_t instead of (int) for vertex pointer offsets (CCSprite/CCTextureAtlas)
- Cast [target rotation] to CCNode* to avoid the new CGVector selector clash
- Install a rootViewController hosting the EAGLView (required since iOS 13)
- Load textures synchronously (background shared EAGLContext is unreliable)
- Disable TextureAtlas VBOs (software ES1 renderer drew batched geometry black)
- Replace FBO-based CCCrossFadeTransition with non-FBO CCFadeTransition
- Defer the intro->game scene switch so the Game scene receives onEnter
- convertToGL: map view points through the GL surface so touches line up

Gameplay:
- Replace the airplane's CCRotateTo turn animation (corrupted by the action
  manager on modern iOS) with a manual shortest-path heading interpolation

+++ Airliner_USA/code/Airliner USA/libs/cocos2d/CCTextureAtlas.m
@@ -311,7 +311,7 @@
 	
 #else // ! CC_TEXTURE_ATLAS_USES_VBO
 	
-	int offset = (int)quads_;
+	intptr_t offset = (intptr_t)quads_;
 
 	// vertex
 	int diff = offsetof( ccV3F_C4B_T2F, vertices);
