+++ Airliner_World/code/Airliner/libs/cocos2d/CCSprite.m
@@ -521,7 +521,7 @@ struct transformValues_ {
 #define kQuadSize sizeof(quad_.bl)
 	glBindTexture(GL_TEXTURE_2D, [texture_ name]);
 	
-	int offset = (int)&quad_;
+	intptr_t offset = (intptr_t)&quad_;
 	
 	// vertex
 	int diff = offsetof( ccV3F_C4B_T2F, vertices);
